examples/gke-sample/requirements.txt
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
google-api-core==0.1.3
google-auth==1.2.1
google-cloud-pubsub==0.30.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.8.2
idna==2.6
protobuf==3.5.1
psutil==5.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
